pyproject.toml
[project]
name = "social-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "fastapi-cache2[redis]>=0.2.2",
    "pydantic[email]>2",
    "pydantic",
    "sqlalchemy>=2.0.38",
    "pydantic-settings==2.7.1",
    "python-jose>=3.4.0",
    "greenlet>=3.1.1",
    "celery>=5.4.0",
    "pillow>=11.1.0",
    "flower>=2.0.1",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
]
